Ingest-api refactors
