+++ .github/workflows/develop.yml
@@ -29,7 +29,7 @@ jobs:
         run: pre-commit run --all-files
     
   test-dev:
-    needs: [lint]
+    # needs: [lint]
     runs-on: ubuntu-latest
     steps:
       - uses: actions/checkout@v3
@@ -62,8 +62,11 @@ jobs:
       - name: Integrations tests
         run: python -m pytest .github/workflows/tests/ -vv -s
       
+      - name: Install reqs for unit tests
+        run: python -m pip install -r ingest_api/requirements.txt
+      
       - name: Ingest unit tests
-        run: python -m pytest ingest_api/tests/ -vv -s
+        run: NO_PYDANTIC_SSM_SETTINGS=1 python -m pytest ingest_api/tests/ -vv -s
 
       - name: Stop services
         run: docker compose stop
