+++ ingest_api/src/schemas.py
@@ -2,10 +2,9 @@ import base64
 import binascii
 import enum
 import json
-import re
 from datetime import datetime
 from decimal import Decimal
-from typing import TYPE_CHECKING, Dict, List, Literal, Optional, Union
+from typing import TYPE_CHECKING, Dict, List, Optional
 from urllib.parse import urlparse
 
 import src.validators as validators
@@ -17,10 +16,9 @@ from pydantic import (
     root_validator,
     validator,
 )
-from src.schema_helpers import BboxExtent, SpatioTemporalExtent, TemporalExtent
+from src.schema_helpers import SpatioTemporalExtent
 from stac_pydantic import Collection, Item, shared
 from stac_pydantic.links import Link
-from typing_extensions import Annotated
 
 from fastapi.exceptions import RequestValidationError
 
@@ -230,156 +228,3 @@ class ListIngestionResponse(BaseModel):
 class UpdateIngestionRequest(BaseModel):
     status: Status = Field(None, description="Status of the ingestion")
     message: str = Field(None, description="Message of the ingestion")
-
-
-class WorkflowInputBase(BaseModel):
-    collection: str = ""
-    upload: Optional[bool] = False
-    cogify: Optional[bool] = False
-    dry_run: bool = False
-
-    @validator("collection")
-    def exists(cls, collection):
-        """
-        Validate that the collection exists.
-
-        Parameters:
-        - collection (str): Name of the collection to be validated.
-
-        Returns:
-        - str: Name of the collection.
-        """
-        validators.collection_exists(collection_id=collection)
-        return collection
-
-
-class S3Input(WorkflowInputBase):
-    discovery: Literal["s3"]
-    prefix: str
-    bucket: str
-    filename_regex: str = r"[\s\S]*"  # default to match all files in prefix
-    datetime_range: Optional[str]
-    start_datetime: Optional[datetime]
-    end_datetime: Optional[datetime]
-    single_datetime: Optional[datetime]
-    zarr_store: Optional[str]
-
-    @root_validator
-    def object_is_accessible(cls, values):
-        bucket = values.get("bucket")
-        prefix = values.get("prefix")
-        zarr_store = values.get("zarr_store")
-        validators.s3_bucket_object_is_accessible(
-            bucket=bucket, prefix=prefix, zarr_store=zarr_store
-        )
-        return values
-
-
-class CmrInput(WorkflowInputBase):
-    discovery: Literal["cmr"]
-    version: Optional[str]
-    include: Optional[str]
-    temporal: Optional[List[datetime]]
-    bounding_box: Optional[List[float]]
-
-
-# allows the construction of models with a list of discriminated unions
-ItemUnion = Annotated[Union[S3Input, CmrInput], Field(discriminator="discovery")]
-
-
-class Dataset(BaseModel):
-    collection: str
-    title: str
-    description: str
-    license: str
-    is_periodic: Optional[bool] = False
-    time_density: Optional[str] = None
-    links: Optional[List[Link]] = []
-    discovery_items: List[ItemUnion]
-
-    # collection id must be all lowercase, with optional - delimiter
-    @validator("collection")
-    def check_id(cls, collection):
-        if not re.match(r"[a-z]+(?:-[a-z]+)*", collection):
-            raise ValueError(
-                "Invalid id - id must be all lowercase, with optional '-' delimiters"
-            )
-        return collection
-
-    @root_validator
-    def check_time_density(cls, values):
-        validators.time_density_is_valid(values["is_periodic"], values["time_density"])
-        return values
-
-
-class DataType(str, enum.Enum):
-    cog = "cog"
-    zarr = "zarr"
-
-
-class COGDataset(Dataset):
-    spatial_extent: BboxExtent
-    temporal_extent: TemporalExtent
-    sample_files: List[str]  # unknown how this will work with CMR
-    data_type: Literal[DataType.cog]
-
-    @root_validator
-    def check_sample_files(cls, values):
-        # pydantic doesn't stop at the first validation,
-        # if the validation for s3 item access fails, "discovery_items" isn't returned
-        # this avoids throwing a KeyError
-        if not (discovery_items := values.get("discovery_items")):
-            return
-
-        if "s3" not in [item.discovery for item in discovery_items]:
-            return values
-
-        # TODO cmr handling/validation
-        invalid_fnames = []
-        for fname in values.get("sample_files", []):
-            found_match = False
-            for item in discovery_items:
-                if all(
-                    [
-                        item.discovery == "s3",
-                        re.search(item.filename_regex, fname.split("/")[-1]),
-                        "/".join(fname.split("/")[3:]).startswith(item.prefix),
-                    ]
-                ):
-                    if item.datetime_range:
-                        try:
-                            validators.extract_dates(fname, item.datetime_range)
-                        except Exception:
-                            raise ValueError(
-                                f"Invalid sample file - {fname} does not align"
-                                "with the provided datetime_range, and a datetime"
-                                "could not be extracted."
-                            )
-                    found_match = True
-            if not found_match:
-                invalid_fnames.append(fname)
-        if invalid_fnames:
-            raise ValueError(
-                f"Invalid sample files - {invalid_fnames} do not match any"
-                "of the provided prefix/filename_regex combinations."
-            )
-        return values
-
-
-class ZarrDataset(Dataset):
-    xarray_kwargs: Optional[Dict] = dict()
-    x_dimension: Optional[str]
-    y_dimension: Optional[str]
-    temporal_dimension: Optional[str]
-    reference_system: Optional[int]
-    data_type: Literal[DataType.zarr]
-
-    @validator("discovery_items")
-    def only_one_discover_item(cls, discovery_items):
-        if len(discovery_items) != 1:
-            raise ValueError("Zarr dataset should have exactly one discovery item")
-        if not discovery_items[0].zarr_store:
-            raise ValueError(
-                "Zarr dataset should include zarr_store in its discovery item"
-            )
-        return discovery_items
